+++ src/features/basic/index.ts
@@ -19,6 +19,7 @@ import './contd-upward-search-results';
 import './cx-search-bar';
 import './cxob-center-on-open';
 import './cxob-supply-demand-values/cxob-supply-demand-values';
+import './cxpc-chart-types/cxpc-chart-types';
 import './cxpo-auto-price';
 import './cxpo-bigger-buttons';
 import './correct-commands/correct-commands';